backend/requirements.txt
# --- Core Backend ---
Flask==3.1.2
Flask-Cors==4.0.0
cachetools
gunicorn==21.2.0
python-dotenv==1.0.1
waitress==3.0.0
beautifulsoup4==4.12.3
cryptography==46.0.4
cycler==0.12.1
Flask-Migrate==4.1.0
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
gast==0.7.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.78.0
h5py==3.15.1
html5lib==1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6

# --- Database ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
Flask-SQLAlchemy
PyJWT==2.8.0
alembic==1.18.3

# --- ML / Statistical Inference ---
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.2
torch==2.2.2

# --- API Client ---
requests==2.32.3
yahooquery
